% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_yarra_app.R
\name{run_yarra_app}
\alias{run_yarra_app}
\title{Launch the Yarra River Shiny app}
\usage{
run_yarra_app(...)
}
\value{
No return value; called for side effects.
}
\description{
Launch the Yarra River Shiny app
}
\examples{
\dontrun{
  launch_app()
}

}
