% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarrariver-package.R
\docType{package}
\name{yarrariver-package}
\alias{yarrariver}
\alias{yarrariver-package}
\title{yarrariver: Explore Yarra River Water Quality}
\description{
An installable R package with a Shiny app and documented datasets to explore Yarra River water quality (1990s vs recent).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-KaiyiYe}
  \item Report bugs at \url{https://github.com/ETC5523-2025/assignment-4-packages-and-shiny-apps-KaiyiYe/issues}
}

}
\author{
\strong{Maintainer}: Kaiyi Ye \email{kyee0020@student.monash.edu}

}
\keyword{internal}
